refactor: extract reusable UI components
